requirements.txt
torch>=2.3.0
transformers>=4.44.0
accelerate>=0.33.0
datasets>=2.20.0
scipy>=1.13.0
numpy>=1.26.0
matplotlib>=3.9.0
pyyaml>=6.0.1
tqdm>=4.66.0
safetensors>=0.4.0
tokenizers>=0.19.0
sentencepiece>=0.2.0
